# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "task-manager-api",
    "description": "任務管理api實作"
  },
  "host": "billhuang-task-manager.herokuapp.com",
  "basePath": "/",
  "tags": [
    {
      "name": "users",
      "description": "user router"
    },
    {
      "name": "tasks",
      "description": "task router"
    }
  ],
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "新增使用者帳號",
        "description": "新增登入用的使用者帳號",
        "parameters": [
          {
            "name": "newUser",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "取得使用者資訊",
        "description": "取得使用者資訊",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "更新使用者資料",
        "description": "更新使用者資料",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "刪除使用者資料",
        "description": "刪除使用者資料",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "登入",
        "description": "輸入帳號密碼後進行登入",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "登出",
        "description": "移除當前使用者所使用的token",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/logoutAll": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "全部登出",
        "description": "移除當前使用者所有可以使用的token",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/me/avatar": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "上傳個人照片",
        "description": "將上傳的圖片檔轉為二進制存入db，限定jpg|jpeg|png",
        "parameters": [
          {
            "name": "avatar",
            "in": "formData",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "刪除個人照片",
        "description": "將使用者圖片檔刪除",
        "parameters": [],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/users/{id}/avatar": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "取得個人圖片",
        "description": "取得db內二進制的圖片",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "新增任務",
        "description": "新增代辦事項的任務",
        "parameters": [
          {
            "name": "task",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/task"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "取得任務",
        "description": "取得此使用者所有代辦事項，並做出分頁及排序或篩選",
        "parameters": [
          {
            "name": "completed",
            "in": "query",
            "type": "string",
            "description": "是否完成-false",
            "required": false
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string",
            "description": "依照什麼做排序-createdAt:desc",
            "required": false
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string",
            "description": "一次請求最多幾筆資料回傳-10",
            "required": false
          },
          {
            "name": "skip",
            "in": "query",
            "type": "string",
            "description": "省略幾筆-0",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "取特定代辦事項",
        "description": "依照task-id取得特定的代辦事項資料",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "tasks"
        ],
        "summary": "取得特定代辦事項",
        "description": "依照task-id取得特定的代辦事項資料",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "刪除特定代辦事項",
        "description": "依照task-id取得特定的代辦事項資料",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "email@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "yourpassword"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "AddUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jahn"
        },
        "email": {
          "type": "string",
          "example": "riwkajv@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "yourpassword"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ]
    },
    "task": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "example": "寫入代辦事項"
        },
        "completed": {
          "type": "boolean",
          "example": false
        }
      },
      "required": [
        "description",
        "completed"
      ]
    }
  }
}